+++ MeshSimplification/geometry/vertex.h
@@ -19,10 +19,10 @@ public:
 	 * \param id The vertex ID.
 	 * \param position The vertex position.
 	 */
-	Vertex(const std::size_t id, const glm::vec3& position) noexcept : id_{id}, position_{position} {}
+	Vertex(const std::uint64_t id, const glm::vec3& position) noexcept : id_{id}, position_{position} {}
 
 	/** \brief Gets the vertex ID. */
-	[[nodiscard]] std::size_t id() const noexcept { return id_; }
+	[[nodiscard]] std::uint64_t id() const noexcept { return id_; }
 
 	/** \brief Gets the vertex position. */
 	[[nodiscard]] const glm::vec3& position() const noexcept { return position_; }
@@ -34,19 +34,19 @@ public:
 	void set_edge(const std::shared_ptr<const HalfEdge>& edge) noexcept { edge_ = edge; }
 
 	/** \brief Gets the hash value for a vertex. */
-	friend std::size_t hash_value(const Vertex& v0) noexcept { return std::hash<std::size_t>{}(v0.id_); }
+	friend std::uint64_t hash_value(const Vertex& v0) noexcept { return std::hash<std::uint64_t>{}(v0.id_); }
 
 	/** \brief Gets the hash value for two vertices. */
-	friend std::size_t hash_value(const Vertex& v0, const Vertex& v1) noexcept {
-		std::size_t seed = 0x32C95994;
+	friend std::uint64_t hash_value(const Vertex& v0, const Vertex& v1) noexcept {
+		std::uint64_t seed = 0x32C95994;
 		seed ^= (seed << 6) + (seed >> 2) + 0x3FA612CE + hash_value(v0);
 		seed ^= (seed << 6) + (seed >> 2) + 0x197685C2 + hash_value(v1);
 		return seed;
 	}
 
 	/** \brief Gets the hash value for three vertices. */
-	friend std::size_t hash_value(const Vertex& v0, const Vertex& v1, const Vertex& v2) noexcept {
-		std::size_t seed = 0x230402B5;
+	friend std::uint64_t hash_value(const Vertex& v0, const Vertex& v1, const Vertex& v2) noexcept {
+		std::uint64_t seed = 0x230402B5;
 		seed ^= (seed << 6) + (seed >> 2) + 0x72C2C6EB + hash_value(v0);
 		seed ^= (seed << 6) + (seed >> 2) + 0x16E199E4 + hash_value(v1);
 		seed ^= (seed << 6) + (seed >> 2) + 0x6F89F2A8 + hash_value(v2);
@@ -54,7 +54,7 @@ public:
 	}
 
 private:
-	std::size_t id_;
+	std::uint64_t id_;
 	glm::vec3 position_;
 	std::weak_ptr<const HalfEdge> edge_;
 };
